First big step migrating the SL into the WebUI done: AccountService is working.
The account service is to create an account, login/logout and get
account details. All the services at AccountManagementRPCImpl now
redirect to the service layer and reuqests the services there. All
persistence lines are deleted in the WebUI related to the account
services.
To store the token for given sessions, a TokenManager has been
introduced. It will replace the old UserManager, as the WebUI does not
have "Users" as such anymore.

To inject the TokenManager, I needed to modify the Security class, which
provides methods to check a given token for authentication. The token
itself is 

The ClientFactory provides a client creation for jax-rs request calls.

The StartUp does neither start the scheduler for experiment scheduling
nor the ServerSocket for MEC connections any more. The serivce layer is
now responsible for these.

The GetRPCImpl was marked with deprecated methods and comments, where
the corresponding services can be found in the SL (the URI).

The ScenarioManagerRPCImpl, ExecuteRPCImpl, ResultRPCImpl got some
comments, which need to be watched out for when migrating these classes
to work with the SL.

The ServerCheck class has been modified, but is not needed anymore. At
the moment it's left there, because the integration process is not at
the ServerCheck at the moment.

Added a lot of comments to the SPCRemoteServlet and refactored the code
to work with the TokenManager now.

Needed to add the dependency for a jersey client (of course) and to the
sopeco service layer in the pom.xml.

Change-Id: If6ba186b0ce05d698bc96d9d66b0fd32f1eb45fd
Signed-off-by: Peter Merkert <p.merkert@sap.com>