Added comments, where NULL pointers appear when running the WebUI.
Added comments, where it's needed to be checked if the methods must call
the RESTful interface to update the data there, too.

Added debug outputs, to have more information about the GWT process when
running the WebUI.

Changed the method to store a scenario, to actually UPDATE it in the
database and not.

Changed a lot of method calls in the SpecificationModul.java from
deprecated methods to the new methods.

Change-Id: Iaeea6f29deac8b0a1d80e5f1494f91985a4bd098
Signed-off-by: d061465 <p.merkert@sap.com>